requirements.txt
pillow
pygame
numpy
opencv-python
pywin32
requests
keyboard
setuptools